chore: Rename files and remove unnecessary files
