+++ earthdev/src/app/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Reception } from "@/components/Reception";
+import { Reception } from "@/components/reception";
 import Link from "next/link";
 import classes from "@/components/classes.module.css";
 
@@ -10,12 +10,10 @@ export default function Home() {
         <Reception />
         <br />
         <Link href="https://github.com/Takoyaki-neonz-net/EarthRadar">
-        <button className={classes.button}>
-          github
-        </button>
+          <button className={classes.button}>github</button>
         </Link>
         <br />
-            {/*
+        {/*
         <Link href="localhost:4140">
         修正予定
           <button className={classes.button}>Go to Home</button>
